pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "hydrobot"
description = "A suite of processing tools for Hilltop hydrological data."
version = "0.3.2"
authors = [
    { name = "Nic Mostert", email = "nicolas.mostert@horizons.govt.nz" },
    { name = "Sam Irvine", email = "sam.irvine@horizons.govt.nz" }
]
requires-python = "==3.11.6"
dependencies = [
    "hilltop-py>=2.3.1",
    "data-annalist>=0.3.6",
    "matplotlib>=3.8.0"
]
classifiers=[
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Natural Language :: English",
    "Programming Language :: Python :: 3.11",
]
license = {text = "GNU General Public License v3"}
readme = "README.rst"
keywords = ["hydrology, automation, pandas, hilltop, hilltop-py, HorizonsRC"]

[project.urls]
Homepage = "https://github.com/HorizonsRC/hydrobot"
Issues = "https://github.com/HorizonsRC/hydrobot/issues"
Documentation = "https://hydrobot.readthedocs.io"
Package = "https://pypi.org/project/hydrobot"


[project.optional-dependencies]
test = [
    "pytest>=7.4.2",
    "pytest-cov>=4.1.0",
    "pytest-dependency>=0.5.1",
    "pytest-mock>=3.12.0",
]
dev = [
    "ruff>=0.1.6",
    "ruff-lsp>=0.0.45",
    "pre-commit>=3.5.0",
]
docs = [
    "Sphinx>=7.2.6",
    "furo>=2023.9.10",
]
all = [
    "hydrobot[test, dev, docs]"
]

[tool.setuptools]
packages = ["hydrobot"]

[tool.setuptools.package-data]
hydrobot = ["hydrobot/config/*"]

[tool.pytest.ini_options]
addopts = "--cov --cov-fail-under 100"

[tool.coverage.run]
source = ["hydrobot"]


[tool.ruff]
extend-exclude = ["docs/*"]
extend-select = [
    "W505", # Line too long as a warning, not an error. Default set to 88.
    "UP",
	"B", # Flake8 Bugbears:
	"PD", # Pandas pro-gamer tips and conventions
	"S", # Flake8 Bandit: Common security flaws.
	"SLF",
	"SIM",
	"I",
	"N",
	"D",
	"PT",
]
extend-ignore = ["D401", "D203", "D212"]

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.ruff.per-file-ignores]
"tests/*" = ["S311", "S101", "F841"]
"docs/*" = ["I001"]
"prototypes" = ["D"]

[tool.bumpversion]
current_version = "0.3.2"
commit = true
tag = true
tag_name = "{new_version}"

[[tool.bumpversion.files]]
filename = "pyproject.toml"
search = "{current_version}"
replace = "{new_version}"

[[tool.bumpversion.files]]
filename = "hydrobot/__init__.py"
search = "__version__ = \"{current_version}\""
replace = "__version__ = \"{new_version}\""
